+++ isaaclab_arena_curobo/ik_reachability_validator.py
@@ -20,8 +20,8 @@ from isaaclab_arena.relations.placement_validation import PlacementCheck, Placem
 from isaaclab_arena.relations.placement_validators import PlacementValidator
 from isaaclab_arena.relations.relations import get_anchor_objects
 from isaaclab_arena.utils.yaw import rotate_quat_by_yaw, yaw_from_quat_xyzw
-from isaaclab_arena_curobo.curobo_ik_utils import (
-    CuroboIKSolver,
+from isaaclab_arena_curobo.ik_solver import CuroboIKSolver
+from isaaclab_arena_curobo.ik_solver_utils import (
     check_ik_feasibility,
     get_aabb_collision_cuboid_for_object,
     top_down_grasp_pose_from_world_poses,
